rename add ride page & submit; fix an import in member_dao
