+++ lib/widgets/pages/rideList/rideListPage.dart
@@ -6,7 +6,7 @@ import 'package:weforza/injection/injectionContainer.dart';
 import 'package:weforza/model/ride.dart';
 import 'package:weforza/repository/ride_repository.dart';
 import 'package:weforza/widgets/common/generic_error.dart';
-import 'package:weforza/widgets/pages/addRide/addRidePage.dart';
+import 'package:weforza/widgets/pages/add_ride/add_ride_page.dart';
 import 'package:weforza/widgets/pages/exportRides/exportRidesPage.dart';
 import 'package:weforza/widgets/pages/rideDetails/rideDetailsPage.dart';
 import 'package:weforza/widgets/pages/rideList/rideListEmpty.dart';
